+++ src/components/Navbar/Navbar.scss
@@ -5,6 +5,10 @@
   display: flex;
 }
 
+.desktop-layout .navbar {
+  transform: translateX(0);
+}
+
 .mobile-layout .navbar {
   display: block;
   position: fixed;
@@ -16,7 +20,12 @@
   height: calc(100vh - #{$header-height});
   overflow-y: auto;
   z-index: 1;
+  transform: translateX(100%);
   transition: transform .2s ease-in-out;
+
+  &.visible {
+    transform: translateX(0);
+  }
 
   > li {
     padding: 10px;
